pyproject.toml
# pyproject.toml

[build-system]
requires = ["setuptools>=69.0.2", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "swanki"
version = "1.0.2"
description = "An example package"
readme = "README.md"
authors = [{ name = "Michael Volk", email = "michaeljvolk7@gmail.com" }]
license = { file = "LICENSE" }
classifiers = [
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3",
]
keywords = ["An example package", "starving artist"]
requires-python = ">=3.13"
dependencies = [
    "python-dotenv",
    "langchain-community",
    "langchain",
    "openai",
    "tiktoken",
    "PyPDF2",
    "hydra-core",
    "omegaconf",
    "pydantic>=2.0",
    "pydantic-ai>=0.1",
    "pyyaml",
    "elevenlabs",
    "pydub",
    "requests",
    "markdown",
    "pyzotero",
]

[project.optional-dependencies]
dev = ["ruff", "semantic-release", "pip-tools", "pytest", "pytest-testmon", "pre-commit", "mypy", "coverage"]

[project.urls]
Homepage = "https://github.com/Mjvolk3/Swanki"

[project.scripts]
swanki = "swanki.__main__:main"
swanki-cut = "swanki.cut:main"
swanki-combine-pdf = "swanki.combine_pdf:main"
swanki-to-anki = "swanki.send_to_anki:main"
swanki-present = "swanki.presentation:main"


[tool.ruff]
target-version = "py313"
line-length = 88

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "D"]
ignore = ["E501"]

[tool.ruff.lint.per-file-ignores]
# Frontmatter docstrings (path, dendron link, github URL) aren't Google-style summaries
"**" = ["D205", "D212", "D415"]
# Test files: pytest hooks and __init__.py don't need docstrings
"tests/**" = ["D100", "D101", "D102", "D103", "D104"]
# Utility scripts: small single-purpose tools, docstrings optional
"notes/assets/scripts/**" = ["D100", "D103"]

[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.ruff.format]
quote-style = "double"

# https://setuptools.pypa.io/en/latest/userguide/pyproject_config.html
[tool.setuptools.dynamic]
dependencies = { file = ["env/requirements.txt"] }
version = { attr = "swanki.__version__" }

[project.entry-points."hydra_plugins"]
swanki = "swanki.config.hydra_plugins"

[tool.setuptools.package-data]
swanki = ["py.typed", "conf/**/*.yaml"]

[tool.setuptools.packages.find]
include = ["swanki*"]

[tool.semantic_release]
version_variables = [
    "pyproject.toml:version",
    "swanki/__version__.py:__version__",
]
branch = "main"
upload_to_pypi = "true"
commit_parser = "scipy"

[tool.semantic_release.commit_parser_options]
allowed_tags = [
    "API",
    "DEP",
    "ENH",
    "REV",
    "BUG",
    "MAINT",
    "BENCH",
    "BLD",
    "DEV",
    "DOC",
    "STY",
    "TST",
    "REL",
    "FEAT",
    "TEST",
]
major_tags = ["API"]
minor_tags = ["DEP", "DEV", "ENH", "REV", "FEAT"]
patch_tags = ["BLD", "BUG", "MAINT"]

[tool.mypy]
python_version = "3.13"
strict = true
plugins = ["pydantic.mypy"]

[[tool.mypy.overrides]]
module = "tests.*"
disallow_untyped_defs = false
disallow_untyped_calls = false
disallow_untyped_decorators = false

[[tool.mypy.overrides]]
module = ["pydub", "pydub.*", "tiktoken", "tiktoken.*", "elevenlabs", "elevenlabs.*", "pydantic_ai", "pydantic_ai.*", "PyPDF2", "PyPDF2.*", "PIL", "PIL.*"]
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = ["swanki.audio.*", "swanki.pipeline.*", "swanki.presentation.*", "swanki.models.*"]
disallow_untyped_calls = false
disable_error_code = ["call-overload", "var-annotated", "type-arg", "unused-ignore", "assignment", "no-any-return", "no-untyped-def"]

[tool.pytest.ini_options]
markers = [
    "llm: LLM API calls, skipped by default (--llm)",
    "integration: integration tests requiring external services, skipped by default (--integration)",
]

[tool.coverage.run]
source = ["swanki"]
branch = true

[tool.coverage.report]
exclude_lines = ["pragma: no cover", "if __name__ == .__main__.", "if TYPE_CHECKING:"]
